pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-bot"
version = "0.1.0"
description = "Production-grade Polymarket Research & Trading Bot"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "openai>=1.12",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "tenacity>=8.2",
    "websockets>=12.0",
    "eval-type-backport>=0.2.0; python_version < '3.10'",
]

[project.optional-dependencies]
clob = ["py-clob-client>=0.1"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
bot = "src.cli:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
